Fix PinPadView Initialization
