+++ WalletWasabi.Gui/Controls/WalletExplorer/SendTabViewModel.cs
@@ -423,7 +423,7 @@ namespace WalletWasabi.Gui.Controls.WalletExplorer
 					PinPadViewModel pinpad = IoC.Get<IShell>().Documents.OfType<PinPadViewModel>().FirstOrDefault();
 					if (pinpad is null)
 					{
-						pinpad = new PinPadViewModel(null);
+						pinpad = new PinPadViewModel(Global);
 						IoC.Get<IShell>().AddOrSelectDocument(pinpad);
 					}
 					var result = await pinpad.ShowDialogAsync();
